+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Screen to view progress for all items that can be donated to Museum. Each item type (bug, fish, art, etc.) shows progress for how many items were found and/or donated.
+- About dialog, showing version number and linking to a list of licenses.
 
 ## [0.3.0] - 2020-07-28
 ### Added